Chore: Update LF tools role
Add GHA workflow to test ansible role
Update tox and pre-commit dependencies

ERROR! the role 'java-install' was not found in <>
The error appears to be in '': line 5, column 7, but may
be elsewhere in the file depending on the exact syntax
problem.

Update role name to match the role path on GHA and
Jenkins jobs.

Ref: https://github.com/ansible/molecule/issues/1042
Change-Id: I0ddf6bc59a54571810e66791aeb11e16fb35710e
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>